+++ docs/changelog.md
@@ -1,4 +1,6 @@
-### Mar 13, 2018 - Version 3.9.6
+### Mar 14, 2018 - Version 3.9.7
+- Added: {user_meta_****} tag to retrieve current logged in user custom meta data
+- Added: {post_meta_****} tag to retrieve current post custom meta data
 - Fix: removed 'wpembed' from tinymce plugin list (was dropped since wordpress 4.8)
 - Added: Option to retrieve current author meta data for dropdown element with
 - Added: `{author_id}` and `{author_name}` tags which do the same thing as the `{post_author_id}` and `{post_author_name}` tags